@@ -28,8 +28,8 @@
         "framer-motion": "^6.2.8",
         "history": "^5.2.0",
         "prettier-plugin-tailwindcss": "^0.1.8",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.1.0",
+        "react-dom": "^18.1.0",
         "react-error-boundary": "^3.1.4",
         "react-hls-player": "^3.0.7",
         "react-i18next": "^11.15.1",
@@ -44,8 +44,8 @@
         "@types/hls.js": "^1.0.0",
         "@types/jest": "^27.4.0",
         "@types/node": "^17.0.8",
-        "@types/react": "^17.0.38",
-        "@types/react-dom": "^17.0.11",
+        "@types/react": "^18.0.12",
+        "@types/react-dom": "^18.0.5",
         "@types/react-router-dom": "^5.3.2",
         "@types/styled-components": "^5.1.24",
         "@typescript-eslint/eslint-plugin": "^5.15.0",
@@ -60,9 +60,9 @@
         "eslint-plugin-prettier": "^4.0.0",
         "eslint-plugin-react": "^7.29.4",
         "eslint-plugin-react-hooks": "^4.3.0",
-        "postcss": "^8.4.6",
+        "postcss": "^8.4.14",
         "prettier": "^2.6.0",
-        "react-scripts": "5.0.0",
+        "react-scripts": "^5.0.1",
         "styled-components": "^5.3.3",
         "tailwindcss": "^3.0.23",
         "typescript": "^4.5.5"
